Move ShouldThrow overloads and their support functions to a dedicated file in anticipation of many type specific overloads.
